wip game layout refactor
