+++ components/level/game.tsx
@@ -1,10 +1,8 @@
 import Direction, { getDirectionFromCode } from '@root/constants/direction';
-import { GameContext } from '@root/contexts/gameContext';
 import { directionsToGameState, isValidDirections } from '@root/helpers/checkpointHelpers';
 import { areEqualGameStates, cloneGameState, GameState, initGameState, makeMove, undo } from '@root/helpers/gameStateHelpers';
 import isPro from '@root/helpers/isPro';
 import useCheckpoints, { BEST_CHECKPOINT_INDEX } from '@root/hooks/useCheckpoints';
-import useDeviceCheck, { ScreenSize } from '@root/hooks/useDeviceCheck';
 import { Types } from 'mongoose';
 import Image from 'next/image';
 import Link from 'next/link';
@@ -16,9 +14,8 @@ import TileType from '../../constants/tileType';
 import { AppContext } from '../../contexts/appContext';
 import { LevelContext } from '../../contexts/levelContext';
 import { PageContext } from '../../contexts/pageContext';
-import Control from '../../models/control';
 import Level, { EnrichedLevel } from '../../models/db/level';
-import GameLayout from './gameLayout';
+import Grid from './grid';
 
 interface SessionCheckpoint {
   _id: Types.ObjectId;
@@ -31,7 +28,6 @@ interface GameProps {
   disablePlayAttempts?: boolean;
   disableStats?: boolean;
   enableSessionCheckpoint?: boolean;
-  extraControls?: Control[];
   level: Level;
   matchId?: string;
   onMove?: (gameState: GameState) => void;
@@ -47,7 +43,6 @@ export default function Game({
   disablePlayAttempts,
   disableStats,
   enableSessionCheckpoint,
-  extraControls,
   level,
   matchId,
   onMove,
@@ -216,32 +211,6 @@ export default function Game({
       }
     });
   }, [gameState, level._id, mutateCheckpoints]);
-
-  const deleteCheckpoint = useCallback((index: number) => {
-    if (index !== BEST_CHECKPOINT_INDEX) {
-      toast.dismiss();
-      toast.loading(`Deleting checkpoint ${index}...`);
-    }
-
-    fetch(`/api/level/${level._id}/checkpoints?index=${index}`, {
-      method: 'DELETE',
-    }).then(async res => {
-      if (res.status === 200) {
-        if (index !== BEST_CHECKPOINT_INDEX) {
-          toast.dismiss();
-          toast.success(`Deleted checkpoint ${index}`);
-        }
-
-        mutateCheckpoints();
-      } else {
-        throw res.text();
-      }
-    }).catch(async err => {
-      console.error(err);
-      toast.dismiss();
-      toast.error(JSON.parse(await err)?.error || 'Error deleting checkpoint');
-    });
-  }, [level._id, mutateCheckpoints]);
 
   const loadCheckpoint = useCallback((index: number) => {
     if (!checkpoints) {
@@ -503,6 +472,8 @@ export default function Game({
 
     const { code } = event;
 
+    console.log('keydown', code);
+
     // prevent arrow keys from scrolling the sidebar
     if (code === 'ArrowUp' || code === 'ArrowDown') {
       event.preventDefault();
@@ -513,6 +484,8 @@ export default function Game({
 
   const handleKeyUpEvent = useCallback((event: KeyboardEvent) => {
     const code = event.code;
+
+    console.log('keyup', code);
 
     if (code.startsWith('Shift')) {
       shiftKeyDown.current = false;
@@ -661,78 +634,6 @@ export default function Game({
     };
   }, [handleBlurEvent, handleKeyDownEvent, handleKeyUpEvent, handleTouchMoveEvent, handleTouchStartEvent, handleTouchEndEvent]);
 
-  const [controls, setControls] = useState<Control[]>([]);
-  const { screenSize } = useDeviceCheck();
-  const isMobile = screenSize < ScreenSize.XL;
-
-  useEffect(() => {
-    const _controls: Control[] = [];
-
-    if (onPrev) {
-      const leftArrow = <svg xmlns='http://www.w3.org/2000/svg' fill='none' viewBox='0 0 24 24' strokeWidth={1.5} stroke='currentColor' className='w-6 h-6'>
-        <path strokeLinecap='round' strokeLinejoin='round' d='M10.5 19.5L3 12m0 0l7.5-7.5M3 12h18' />
-      </svg>;
-      const prevTxt = isMobile ? leftArrow : <div><span className='underline'>P</span>rev Level</div>;
-
-      _controls.push(new Control('btn-prev', () => onPrev(), prevTxt ));
-    }
-
-    const restartIcon = (<svg xmlns='http://www.w3.org/2000/svg' fill='none' viewBox='0 0 24 24' strokeWidth={1.5} stroke='currentColor' className='w-6 h-6'>
-      <path strokeLinecap='round' strokeLinejoin='round' d='M16.023 9.348h4.992v-.001M2.985 19.644v-4.992m0 0h4.992m-4.993 0l3.181 3.183a8.25 8.25 0 0013.803-3.7M4.031 9.865a8.25 8.25 0 0113.803-3.7l3.181 3.182m0-4.991v4.99' />
-    </svg>);
-    const restartTxt = isMobile ? restartIcon : <div><span className='underline'>R</span>estart</div>;
-
-    const undoIcon = (<svg xmlns='http://www.w3.org/2000/svg' fill='none' viewBox='0 0 24 24' strokeWidth={1.5} stroke='currentColor' className='w-6 h-6'>
-      <path strokeLinecap='round' strokeLinejoin='round' d='M9 15L3 9m0 0l6-6M3 9h12a6 6 0 010 12h-3' />
-    </svg>);
-    const undoTxt = isMobile ? undoIcon : <div><span className='underline'>U</span>ndo</div>;
-
-    const redoIcon = (<svg xmlns='http://www.w3.org/2000/svg' fill='none' viewBox='0 0 24 24' strokeWidth={1.5} stroke='currentColor' className='w-6 h-6'>
-      <path strokeLinecap='round' strokeLinejoin='round' d='M15 15l6-6m0 0l-6-6m6 6H9a6 6 0 000 12h3' />
-    </svg>);
-
-    const redoTxt = isMobile ? redoIcon : <div>Redo (<span className='underline'>Y</span>)</div>;
-
-    _controls.push(
-      new Control('btn-restart', () => handleKeyDown('KeyR'), restartTxt),
-      new Control('btn-undo', () => handleKeyDown('Backspace'), undoTxt, false, false, () => {
-        handleKeyDown('Backspace');
-
-        return true;
-      }),
-      new Control(
-        'btn-redo',
-        () => handleKeyDown('KeyY'),
-        <span className='flex gap-2 justify-center select-none'>
-          {!pro && <Image className='pointer-events-none z-0' alt='pro' src='/pro.svg' width='16' height='16' />}
-          {redoTxt}
-        </span>,
-        gameState.redoStack.length === 0,
-        false,
-        () => {
-          handleKeyDown('KeyY');
-
-          return true;
-        },
-      ),
-    );
-
-    if (onNext) {
-      const rightArrow = <span className='truncate'><svg xmlns='http://www.w3.org/2000/svg' fill='none' viewBox='0 0 24 24' strokeWidth={1.5} stroke='currentColor' className='w-6 h-6'>
-        <path strokeLinecap='round' strokeLinejoin='round' d='M13.5 4.5L21 12m0 0l-7.5 7.5M21 12H3' />
-      </svg></span>;
-      const nextTxt = isMobile ? rightArrow : <div><span className='underline'>N</span>ext Level</div>;
-
-      _controls.push(new Control('btn-next', () => onNext(), nextTxt));
-    }
-
-    if (extraControls) {
-      setControls(_controls.concat(extraControls));
-    } else {
-      setControls(_controls);
-    }
-  }, [extraControls, gameState.redoStack.length, handleKeyDown, isMobile, onNext, onPrev, pro, setControls]);
-
   function onCellClick(x: number, y: number) {
     if (isSwiping.current) {
       return;
@@ -748,21 +649,15 @@ export default function Game({
   }
 
   return (
-    <GameContext.Provider value={{
-      checkpoints: checkpoints,
-      deleteCheckpoint: deleteCheckpoint,
-      level: level,
-      loadCheckpoint: loadCheckpoint,
-      saveCheckpoint: saveCheckpoint,
-    }}>
-      <GameLayout
-        controls={controls}
-        disableCheckpoints={disableCheckpoints}
-        gameState={gameState}
-        level={level}
-        matchId={matchId}
-        onCellClick={(x, y) => onCellClick(x, y)}
-      />
-    </GameContext.Provider>
+    <Grid
+      gameState={gameState}
+      id={level._id.toString()}
+      leastMoves={level.leastMoves}
+      onCellClick={(x, y, rightClick) => {
+        if (!rightClick) {
+          onCellClick(x, y);
+        }
+      }}
+    />
   );
 }
