+++ contexts/gameContext.ts
@@ -1,18 +1,15 @@
 import Direction from '@root/constants/direction';
 import Level from '@root/models/db/level';
 import { createContext } from 'react';
+import { KeyedMutator } from 'swr';
 
 interface GameContextInterface {
   checkpoints?: (Direction[] | null)[];
-  deleteCheckpoint: (index: number) => void;
   level: Level;
-  loadCheckpoint: (index: number) => void;
-  saveCheckpoint: (index: number) => void;
+  mutateCheckpoints: KeyedMutator<(Direction[] | null)[]>;
 }
 
 export const GameContext = createContext<GameContextInterface>({
-  deleteCheckpoint: () => { return; },
   level: {} as Level,
-  loadCheckpoint: () => { return; },
-  saveCheckpoint: () => { return; },
+  mutateCheckpoints: {} as KeyedMutator<(Direction[] | null)[]>,
 });
